attempt to fix a crash at runtime when first accessing `paper`
the reason for the crash might have been that during setup of the environment (install Rust and dependencies) a version of `swift-package` was installed which was incompatible with the generated uniffi interface of the paper Swift Package.

The fix attempt is to install the exact version of swift-package that matches the version of uniffi.